fix compilation errors

+++ beacon-chain/rpc/beacon/validators.go
@@ -263,7 +263,7 @@ func (bs *Server) ListValidators(
 	})
 
 	if len(req.PublicKeys) == 0 && len(req.Indices) == 0 {
-		for i := 0; i < reqState.NumValidators(); i++ {
+		for i := uint64(0); i < uint64(reqState.NumValidators()); i++ {
 			val, err := reqState.ValidatorAtIndex(uint64(i))
 			if err != nil {
 				return nil, status.Errorf(codes.Internal, "Could not get validator: %v", err)
